Remove tasks/reply endpoint and task_actions.go
